Add files via upload
We run the W3 model to provide hydrological variables at the resolution of 1°×1°. It covers the period from Jan. 2000 to Dec. 2018, and forced by the ERA5 meteorological data. Additionally, ,we have been trying to improve the temporal and spatial resolution of these products, and achived reasonable improvements. If you need the products with higher resolution (e.g., 0.25° and daily), please contact us by email or wechat (wjyin1991@163.com, BNU18813141283).